backend/audio-service/requirements.txt
asgiref==3.8.1
boto3==1.37.17
botocore==1.37.17
Django==5.1.7
django-cors-headers==4.7.0
django-storages==1.14.5
djangorestframework==3.15.2
jmespath==1.0.1
psycopg2==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
